fix(frontend): Switch Report List view for generic list component
I've created a generic list component, that can be configured with
different list items. This new component just manages the pagination.
This means the index file needs to provide all data.
